graph, server: Split up server modules
