+++ graph/src/components/server/mod.rs
@@ -1,0 +1,8 @@
+/// Component for running GraphQL queries over HTTP.
+pub mod query;
+
+/// Component for running GraphQL subscriptions over WebSockets.
+pub mod subscription;
+
+/// Component for the JSON-RPC admin API.
+pub mod admin;
