+++ graph/src/components/server/subscription.rs
@@ -1,0 +1,1 @@
+
